proyectodw/pyproject.toml
[project]
name = "proyectodw"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.2",
    "pymongo>=4.15.5",
    "python-dotenv>=1.2.1",
]

[tool.uv.workspace]
members = [
    "proyectodw",
]
